requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
numpy==2.4.3
scipy==1.17.1
